pyproject.toml
[project]
name = "andersentesttask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "cryptography>=44.0.3",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"


